+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Fixed
+
+- Fix administrators losing access to a block's configuration after setting a profile to "no access" on that block.
+
 ## [1.24.4] - 2026-08-06
 
 ### Fixed